feat: #tno-2329 - moved Topic component into shared location - updated Event of the day page and Content page top use shared component
