+++ app/editor/src/components/topic/styled/index.ts
@@ -1,0 +1,1 @@
+export * from './Topic';
