+++ app/editor/src/features/content/form/components/index.ts
@@ -1,5 +1,5 @@
+export * from '../../../../components/topic';
 export * from './tags';
 export * from './time-log';
 export * from './tool-bar';
-export * from './topic';
 export * from './upload';
