+++ app/editor/src/features/content/form/interfaces/IGroupedTopicOptions.tsx
@@ -1,0 +1,6 @@
+import { ITopicOptionItem } from './ITopicOptionItem';
+
+export interface IGroupedTopicOptions {
+  readonly label: string;
+  readonly options: readonly ITopicOptionItem[];
+}
