+++ app/editor/src/features/content/form/interfaces/ITopicOptionItem.tsx
@@ -1,0 +1,8 @@
+import { TopicTypeName } from 'tno-core';
+
+export interface ITopicOptionItem {
+  value: number;
+  label: string;
+  topicType: TopicTypeName;
+  isDisabled: boolean;
+}
