+++ app/editor/src/features/content/form/interfaces/index.ts
@@ -1,1 +1,3 @@
 export * from './IContentForm';
+export * from './IGroupedTopicOptions';
+export * from './ITopicOptionItem';
