+++ app/editor/src/features/content/form/styled/ContentForm.tsx
@@ -174,4 +174,8 @@ export const ContentForm = styled.div`
   .tab-layout {
     width: 65%;
   }
+
+  .topic-select {
+    width: 30ch;
+  }
 `;
